requirements.txt
pillow
pip
setuptools
torch
torchvision
wheel